refactor(submissions): turn page component to ssr and extract submissions list into a separate component
